# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aldis API",
        "description": "API documentation for Aldis application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/product-types": {
            "get": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Get all product types",
                "description": "Retrieve a list of all product types.",
                "operationId": "c58762a61e2b123c2028387c49364a6c",
                "responses": {
                    "200": {
                        "description": "List of product types",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Create a new product type",
                "description": "Add a new product type to the system.",
                "operationId": "ee8038a52e0d1a876aaa6f775037bde1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Electronics"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product type created successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/product-types/{id}": {
            "get": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Get a product type by ID",
                "description": "Retrieve details of a specific product type by its ID.",
                "operationId": "3641c650e3e0b0c0a6479ad8cab8c786",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product type ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product type details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product type not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Update a product type",
                "description": "Update details of an existing product type.",
                "operationId": "df5c3f216c562f16afb557fd0ffabbba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product type ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Electronics"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Updated description"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary",
                                        "example": "image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product type updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product type updated successfully"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product type not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product Types"
                ],
                "summary": "Delete a product type",
                "description": "Remove a product type from the system.",
                "operationId": "de414ced497f2688fb9b26cbfa7b26a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product type ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product type deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Product type deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product type not found"
                    }
                }
            }
        },
...
    },
...
}